Distress Signal Unit For FireFighters
More often than not we find that there are many instances

where firefighters go missing during the operations which is due to communication failure. In order to address the issue, we have developed a distress signal unit which is an automatic emergency triggering system, which constantly monitors your movement with the help of gyroscope and uses a combination of GPS and GSM to notify others. The gyro sensor has been programmed in such a way that, if the movement is not detected for 60 seconds or if there is a sudden drastic change in the coordinates of your position then the system interprets that there is some problem. In order to notify others the GPS module Neo-6m determines the location coordinates for that position which is then communicated with the help of GSM module.

The GSM module actually sends the message to the desired number as a hyper link which redirects you to the google maps navigation. This makes it very easy to track the person. Sensors used: MPU6050 (Gyro), Neo6m (GPS) and SIM800C (GSM)